requirements.txt
jupyter
langchain
pypdf
langchain-openai
langchain_community
python-dotenv
pillow # PyPDFLoader
faiss-cpu